final result

+++ src/hooks/useNoteOptions.ts
@@ -1,9 +1,15 @@
 import { TNoteUseOptions } from "../types/types";
 import { db } from "../firebase/firebase";
 import { deleteDoc, doc, getDoc, updateDoc } from "firebase/firestore";
+import { useAppSelector } from "../app/hooks";
+import { selectUserInfo } from "../features/userInfoSlice";
 export default function useNoteOptions(): TNoteUseOptions {
+  const userInfo = useAppSelector(selectUserInfo);
   const archiveNote = async (id: string): Promise<void> => {
-    const docRef = doc(db, "notes", id);
+    const userId = userInfo?.uid;
+    if (!userId) return;
+    const userRef = doc(db, "users", userId);
+    const docRef = doc(userRef, "notes", id);
     await updateDoc(docRef, { isArchived: true });
     const curNoteInfo = await getNoteById(id);
     if (curNoteInfo?.isPinned) {
@@ -11,11 +17,17 @@ export default function useNoteOptions(): TNoteUseOptions {
     }
   };
   const unArchiveNote = async (id: string): Promise<void> => {
-    const docRef = doc(db, "notes", id);
+    const userId = userInfo?.uid;
+    if (!userId) return;
+    const userRef = doc(db, "users", userId);
+    const docRef = doc(userRef, "notes", id);
     await updateDoc(docRef, { isArchived: false });
   };
   const trashNote = async (id: string): Promise<void> => {
-    const docRef = doc(db, "notes", id);
+    const userId = userInfo?.uid;
+    if (!userId) return;
+    const userRef = doc(db, "users", userId);
+    const docRef = doc(userRef, "notes", id);
     await updateDoc(docRef, { isTrashed: true });
     const curNoteInfo = await getNoteById(id);
     if (curNoteInfo?.isPinned) {
@@ -23,30 +35,42 @@ export default function useNoteOptions(): TNoteUseOptions {
     }
   };
   const unTrashNote = async (id: string): Promise<void> => {
-    const docRef = doc(db, "notes", id);
+    const userId = userInfo?.uid;
+    if (!userId) return;
+    const userRef = doc(db, "users", userId);
+    const docRef = doc(userRef, "notes", id);
     await updateDoc(docRef, { isTrashed: false });
   };
 
   const deleteNote = async (id: string): Promise<void> => {
-    const docRef = doc(db, "notes", id);
+    const userId = userInfo?.uid;
+    if (!userId) return;
+    const userRef = doc(db, "users", userId);
+    const docRef = doc(userRef, "notes", id);
     await deleteDoc(docRef);
   };
   const changeNoteBackground = async (
     id: string | string[],
     newColor: string
   ): Promise<void> => {
+    const userId = userInfo?.uid;
+    if (!userId) return;
+    const userRef = doc(db, "users", userId);
     if (!Array.isArray(id)) {
-      const docRef = doc(db, "notes", id);
+      const docRef = doc(userRef, "notes", id);
       await updateDoc(docRef, { noteBackgroundColor: newColor });
     } else {
       id.forEach(async (noteId: string) => {
-        const docRef = doc(db, "notes", noteId);
+        const docRef = doc(userRef, "notes", noteId);
         await updateDoc(docRef, { noteBackgroundColor: newColor });
       });
     }
   };
   const pinAndUnpinNote = async (id: string): Promise<void> => {
-    const docRef = doc(db, "notes", id);
+    const userId = userInfo?.uid;
+    if (!userId) return;
+    const userRef = doc(db, "users", userId);
+    const docRef = doc(userRef, "notes", id);
     let curNoteInfo = await getNoteById(id);
     await updateDoc(docRef, { isPinned: !curNoteInfo?.isPinned });
     curNoteInfo = await getNoteById(id);
@@ -55,7 +79,10 @@ export default function useNoteOptions(): TNoteUseOptions {
     }
   };
   const getNoteById = async (id: string) => {
-    const docRef = doc(db, "notes", id);
+    const userId = userInfo?.uid;
+    if (!userId) return;
+    const userRef = doc(db, "users", userId);
+    const docRef = doc(userRef, "notes", id);
     const docSnap = await getDoc(docRef);
     return docSnap.data();
   };
@@ -63,7 +90,10 @@ export default function useNoteOptions(): TNoteUseOptions {
     newValues: { noteTitle: string; noteValue: string },
     id: string
   ): Promise<void> => {
-    const docRef = doc(db, "notes", id);
+    const userId = userInfo?.uid;
+    if (!userId) return;
+    const userRef = doc(db, "users", userId);
+    const docRef = doc(userRef, "notes", id);
     await updateDoc(docRef, newValues);
   };
   return {
